slider updated
